Avoid writing huge stripes due to direct encode conversion
We always convert column to direct encoding when stripe is small after
05b66c041970fff317f88b15bbc991bd8d1ead35. However, in cases that
columns has extremely well compression, it can end up with writing
huge stripes (larger than 2G in a corner case that each value is the
same array of varchar).

This commit fixed this by only convert to direct encode when the new
size is under a given limit.